libs restructure for better management
