+++ src/AdvertisedData.js
@@ -9,9 +9,9 @@ import {
   useToast, useDisclosure, 
 } from '@chakra-ui/react';
 import ShortAddress from './ShortAddress';
-import { TERMS, CHAIN_TOKEN_SYMBOL } from './util';
-import { config } from './util';
-import { ChainMetaContext } from './contexts';
+import { TERMS, CHAIN_TOKEN_SYMBOL } from './libs/util';
+import { config } from './libs/util';
+import { ChainMetaContext } from './libs/contexts';
 
 export default function() {
   const chainMeta = useContext(ChainMetaContext);
