+++ src/App.js
@@ -28,9 +28,9 @@ import DataStreams from './DataStreams';
 import DataCoalitions from './DataCoalitions';
 import DataCoalitionsViewAll from './DataCoalition/DataCoalitionsViewAll';
 import TrustedComputation from './TrustedComputation';
-import { sleep, contractsForChain, noChainSupport } from './util';
-import { MENU, ABIS, CHAINS, SUPPORTED_CHAINS, CHAIN_TOKEN_SYMBOL } from './util';
-import { chainMeta, ChainMetaContext } from './contexts';
+import { sleep, contractsForChain, noChainSupport } from './libs/util';
+import { MENU, ABIS, CHAINS, SUPPORTED_CHAINS, CHAIN_TOKEN_SYMBOL } from './libs/util';
+import { chainMeta, ChainMetaContext } from './libs/contexts';
 import logo from './img/logo.png';
 import logoSml from './img/logo-sml.png';
 import chainEth from './img/eth-chain-logo.png';
