+++ src/DataNFT/DataNFTMarketplace.js
@@ -8,9 +8,8 @@ import {
 } from '@chakra-ui/react';
 import { ExternalLinkIcon } from '@chakra-ui/icons';
 import ShortAddress from '../ShortAddress';
-import { TERMS, CHAIN_TOKEN_SYMBOL, OPENSEA_CHAIN_NAMES, CHAIN_NAMES, CHAIN_TX_VIEWER, CHAINS } from '../util';
-import { config, mydaRoundUtil } from '../util';
-import { ChainMetaContext } from '../contexts';
+import { TERMS, CHAIN_TOKEN_SYMBOL, OPENSEA_CHAIN_NAMES, CHAIN_NAMES, CHAIN_TX_VIEWER, CHAINS } from '../libs/util';
+import { ChainMetaContext } from '../libs/contexts';
 
 export default function() {
   const chainMeta = useContext(ChainMetaContext);
