+++ src/DataVault.js
@@ -7,8 +7,8 @@ import {
   Text, HStack, Heading, Center,
   useToast
 } from '@chakra-ui/react';
-import { CHAIN_TOKEN_SYMBOL } from './util';
-import { ChainMetaContext } from './contexts';
+import { CHAIN_TOKEN_SYMBOL } from './libs/util';
+import { ChainMetaContext } from './libs/contexts';
 import dataVaultIcon from './img/data-vault-icon.png';
 
 export default function() {
